Revert "🧼 Move polls to root (#5)"
This reverts commit 0855c1b36b360fa5d23578b0d9a6f0175ad28648.
