+++ apps/fc-polls/app/actions.tsx
@@ -3,7 +3,7 @@
 import { kv } from "@vercel/kv";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
-import { Poll, POLL_EXPIRY } from "./polls/types";
+import { Poll, POLL_EXPIRY } from "./types";
 
 export async function savePoll(poll: Poll, formData: FormData) {
   let newPoll = {
