+++ apps/fc-polls/app/favicon.ico
