+++ apps/fc-polls/package.json
@@ -1,0 +1,37 @@
+{
+  "private": true,
+  "scripts": {
+    "dev": "next dev --turbo",
+    "lint": "next lint",
+    "build": "next build",
+    "start": "next start",
+    "test": "concurrently \"npm:lint\" \"npm:build\""
+  },
+  "dependencies": {
+    "@farcaster/hub-nodejs": "^0.11.11",
+    "@types/node": "20.11.0",
+    "@types/react": "18.2.47",
+    "@types/react-dom": "18.2.18",
+    "@types/uuid": "9.0.7",
+    "@vercel/kv": "^1.0.1",
+    "autoprefixer": "10.4.16",
+    "clsx": "^2.1.0",
+    "geist": "^1.2.1",
+    "next": "latest",
+    "postcss": "^8.4.33",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "satori": "^0.10.11",
+    "sharp": "^0.33.2",
+    "tailwindcss": "3.4.1",
+    "typescript": "5.3.3",
+    "uuid": "^9.0.1",
+    "zod": "^3.23.8"
+  },
+  "devDependencies": {
+    "concurrently": "^8.2.2",
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.2.3",
+    "typescript-eslint": "^7.8.0"
+  }
+}
