+++ apps/fc-polls/pages/api/image.tsx
@@ -1,4 +1,4 @@
-import { Poll } from "@/app/polls/types";
+import { Poll } from "@/app/types";
 import { kv } from "@vercel/kv";
 import * as fs from "fs";
 import type { NextApiRequest, NextApiResponse } from "next";
