+++ apps/fc-polls/pages/api/vote.ts
@@ -1,9 +1,9 @@
-import { POLL_EXPIRY, Poll } from "@/app/polls/types";
+import { POLL_EXPIRY, Poll } from "@/app/types";
 import { Message, getSSLHubRpcClient } from "@farcaster/hub-nodejs";
 import { kv } from "@vercel/kv";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { z } from "zod";
-import { env } from "../../../app/env";
+import { env } from "../../app/env";
 
 const client = getSSLHubRpcClient(env.HUB_URL);
 
@@ -28,6 +28,7 @@ export default async function handler(
           Buffer.from(req.body?.trustedData?.messageBytes || "", "hex"),
         );
         const result = await client.validateMessage(frameMessage);
+        console.log({ result });
         if (result && result.isOk() && result.value.valid) {
           validatedMessage = result.value.message;
         }
@@ -54,8 +55,10 @@ export default async function handler(
         buttonId = z.number().parse(req.body?.untrustedData?.buttonIndex);
         fid = z.number().parse(req.body?.untrustedData?.fid);
       }
+      console.log({ buttonId, fid });
 
       // Clicked create poll
+      console.log({ voted, results, buttonId });
       if ((results || voted) && buttonId === 2) {
         return res
           .status(302)
@@ -80,7 +83,7 @@ export default async function handler(
       if (!poll) {
         return res.status(400).send("Missing poll ID");
       }
-      const imageUrl = `${env.HOST}/api/polls/image?id=${poll.id}&results=${
+      const imageUrl = `${env.HOST}/api/image?id=${poll.id}&results=${
         results ? "false" : "true"
       }&date=${Date.now()}${fid > 0 ? `&fid=${fid}` : ""}`;
       let button1Text = "View Results";
@@ -103,7 +106,7 @@ export default async function handler(
           <meta property="og:image" content="${imageUrl}">
           <meta name="fc:frame" content="vNext">
           <meta name="fc:frame:image" content="${imageUrl}">
-          <meta name="fc:frame:post_url" content="${env.HOST}/api/polls/vote?id=${
+          <meta name="fc:frame:post_url" content="${env.HOST}/api/vote?id=${
             poll.id
           }&voted=true&results=${results ? "false" : "true"}">
           <meta name="fc:frame:button:1" content="${button1Text}">
